api refactor for signup the user
